requirements.txt
mysql.connector
streamlit
pandas
XlsxWriter